requirements.txt
python-dotenv
python-escpos
pyusb
requests
